move UpdateSceneWorker to file
